(mostly) got tcpserver reading
I think that it should be able to connect & read (and send events).
Some of the API is iffy, and crashes sometimes, but it *does* read, and
prints (semi) useful messages.
Also, I documented a bit of stuff.
ALSO, I can actually run this thing, which is a plus (and without
eclipse giving me TOO many errors).
...and I probably did a million other amazing things that you could look
through the code for if you *really* need to know, but I'm tired, it's
1:00 AM, and I, for some reason, need sleep. Which is why I don't
remeber everything that I did. But I know that all my changes were
amazing. I think that one of the things had to do with a major
architecture shift, but I'm really too tired to write about it, although
it solved a bunch of big problems that I was having (but might make a
few more, especially when future-me tries to add support for chunked
HTTP requests, unless future-me is too tired or doesn't want to, which
is fine, I guess).
Wow, that was long. Imma sleep now.