Rewrote swaths of main API
Server now extends RunnableService. This allows for stateful servers,
and fixes a lot of things.

Codec has been deprecated, because it's impossible to implement.

Channel was documented, and changed a bit: there is now an option
system, where channels can be marked with data (such as marking a HTTPS
channel as such), without requiring subclassing. Not sure how well this
will work, but let's see!