Implemented HTTP
Not as nice as my TcpServerImpl, but it works 90% of the time. Probably
because HTTP is a 20-year-old protocol, and has become somewhat bloated
with features. I mean, do we *really* need chunked transfers, or
multiple ways to upgrade protocols? Anyway, it works OK, and I'll
probably make API changes soon to fix some of the larger problems (such
as, you can't get the raw data from a HttpRequest).