Started working on auto-updates
It turns out that life is hard. First I have to check whether a new
version has been released (with GitHub's API), then I have to clone it
(with Git), apply preprocessor directives, swap it with the current
resources, clean up the old ones, and finally gradually minify &
compress all the resources, possibly for different browsers. All while
the server is running. And I don't have a good resource API yet.

I also tried to add testing, but it turns out that that takes a *lot* of
time, especially when your spec keeps on changing. Maybe for later.