Bugfix (MAJOR)
There was a bug that would hang any thread that called
TcpClientChannel#connect would hang, because a selector can *only* have
a socket registered on the same thread as the one calling `select`, or
it blocks until select is done (which could take literally forever).

This new method queues an action, then notifies the selector, fixing the
deadlock.

This problem never occurred in TcpServerImpl, because everything that
uses the selector (except for writing data) is on the same thread.