Added ByteBufferPool
Admittedly not finished, but it will hopefully save the GC some work,
because the server will (probably) use similarily-sized ByteBuffers, and
pooling them might even let me make them all direct, or something.
I also should make the test a bit better.
And maybe some histogram thing, so it can retain commonly-used
ByteBuffers (or do some sort of upgrade-to-DirectBuffer thing), while
releasing less-used ones.
And maybe a better lookup thing. I mean, a HashMap for an integer? That
doesn't really make sense, and I'm 90% sure that I could make a
faster-looking-up one. But it would require a lot of work, and would be
complicated. Maybe later (TODO make an issue for this).